requirements.txt
asgiref==3.3.4
Django==3.2
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
gunicorn==20.1.0
setuptools==65.5.1
djangorestframework-simplejwt==4.6.0
Pillow==9.5.0
drf-spectacular